moving options components into option-fields
