+++ src/components/fields/base/option-fields/options/delete.vue
